scripts/requirements.txt
paho-mqtt>=1.6.1,<2
